requirements.txt
# Core ML Libraries
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.35.0
accelerate>=0.24.0
peft>=0.6.0

# Image Processing & OCR
Pillow>=10.0.0
opencv-python>=4.8.0
pytesseract>=0.3.10
easyocr>=1.7.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Metrics and Evaluation
nltk>=3.8.0
rouge-score>=0.1.2

# Demo and UI
gradio>=4.0.0

# Utilities
tqdm>=4.65.0
python-dotenv>=1.0.0
